requirements.txt
anthropic
google-generativeai
llama-index
PyMuPDF
python-dotenv
tree-sitter-language-pack 